refactor: replace direct exec calls with safeExec and move isVersionGte to utils
